feat: Added ProductDTO validation upgrade
https://track.ratio.digital/youtrack/issue/DKCKZ-357
