+++ backend/src/Dto/Product/ProductFilterInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Dto\Product;
+
+interface ProductFilterInterface
+{
+    public function getCode(): string;
+
+    public function getValue(): string;
+
+    public function getUnit(): string;
+}
